Issue 26 update to new context api (#38)
This fix Closes #26

Issue: StoreProvider and Connect used the legacy context API, this change updates it to the new API

Fix: Added a file StoreContext.lua that returns a static Roact Context. StoreProvider now returns a Roact Provider that passes a Rodux Store as its "value". Connect now uses a Roact Consumer to read in the store via props.