LUAFDN-250 - Add types for connect and dispatch prop (#68)
This PR adds types for the dispatch prop and connect. This includes a standard action dispatch and a thunkful action dispatch variant.

Note that I used the thunkful action dispatch as the default for connect(), which may or may not be desirable, but felt like a better default to me.

Inspired by: https://github.com/reduxjs/react-redux/blob/master/src/types.ts

This RP also adds typechecking to CI with luau-lsp.